third_party/blink/web_tests/http/tests/security/cross-frame-access-get-custom-property-cached-expected.txt

This test checks that caching of property accesses doesn't allow for illegal cross-frame access of a custom property. It also checks for an assertion failure that once occurred in this situation.

PASS: customPropertyAccessThrowsException(targetWindow) should be 'true' and is.
PASS: customPropertyAccessThrowsException(targetWindow) should be 'true' and is.
PASS: customPropertyAccessThrowsException(targetWindow) should be 'true' and is.
